+++ src/index.ts
@@ -977,10 +977,32 @@ export async function getTotalDiskCapacity() {
   return -1;
 }
 
+export async function getTotalDiskCapacityNew() {
+  if (Platform.OS === 'android') {
+    return RNDeviceInfo.getTotalDiskCapacityNew();
+  }
+  if (Platform.OS === 'ios' || Platform.OS === 'web') {
+    return getTotalDiskCapacity();
+  }
+
+  return -1;
+}
+
 export function getTotalDiskCapacitySync() {
   if (Platform.OS === 'android' || Platform.OS === 'ios' || Platform.OS === 'web') {
     return RNDeviceInfo.getTotalDiskCapacitySync();
   }
+  return -1;
+}
+
+export function getTotalDiskCapacityNewSync() {
+  if (Platform.OS === 'android') {
+    return RNDeviceInfo.getTotalDiskCapacityNewSync();
+  }
+  if (Platform.OS === 'ios' || Platform.OS === 'web') {
+    return getTotalDiskCapacitySync();
+  }
+
   return -1;
 }
 
@@ -991,10 +1013,32 @@ export async function getFreeDiskStorage() {
   return -1;
 }
 
+export async function getFreeDiskStorageNew() {
+  if (Platform.OS === 'android') {
+    return RNDeviceInfo.getFreeDiskStorageNew();
+  }
+  if (Platform.OS === 'ios' || Platform.OS === 'web') {
+    return getFreeDiskStorage();
+  }
+
+  return -1;
+}
+
 export function getFreeDiskStorageSync() {
   if (Platform.OS === 'android' || Platform.OS === 'ios' || Platform.OS === 'web') {
     return RNDeviceInfo.getFreeDiskStorageSync();
   }
+  return -1;
+}
+
+export function getFreeDiskStorageNewSync() {
+  if (Platform.OS === 'android') {
+    return RNDeviceInfo.getFreeDiskStorageNewSync();
+  }
+  if (Platform.OS === 'ios' || Platform.OS === 'web') {
+    return getFreeDiskStorage();
+  }
+
   return -1;
 }
 
@@ -1376,7 +1420,9 @@ const deviceInfoModule: DeviceInfoModule = {
   getFontScale,
   getFontScaleSync,
   getFreeDiskStorage,
+  getFreeDiskStorageNew,
   getFreeDiskStorageSync,
+  getFreeDiskStorageNewSync,
   getHardware,
   getHardwareSync,
   getHost,
@@ -1420,7 +1466,9 @@ const deviceInfoModule: DeviceInfoModule = {
   getTags,
   getTagsSync,
   getTotalDiskCapacity,
+  getTotalDiskCapacityNew,
   getTotalDiskCapacitySync,
+  getTotalDiskCapacityNewSync,
   getTotalMemory,
   getTotalMemorySync,
   getType,